Problem: building and running complex systems
Modern software systems are typically complex and present significant overhead.
On the architecture spectrum from monolith to SOA, this complexity manifests
through a host of issues.

For example, the data is often copied and transferred over the network to reach
the code that can process it. A simple example would be an application server
sending a query to the database server, getting results, and making further
changes in the database. A more complex example would involve extending this
pipeline with a search engine like Elasticsearch and shuttling data between the
database server and the search engine to get the data indexed. Finally, a
typical example in an SOA architecture would be a service sending data to
another service for further processing; this is particularly common in polyglot
environments where direct integration of service components is less feasible.

Relative closedness of ecosystems of individual stacks results in significant
redundancies and inefficiencies. For example, NLP analysis in Elasticsearch
can't be easily substituted with the one in spaCy (a Python-based NLP toolkit),
so often, the efforts to maintain coherent NLP across the ML and indexing
results in doing the work multiple times, both on the development side as well
as processing.

Another manifestation of complexity comes from observing the management of
system change rollouts. Having multiple components with their own way of
operational management, requirements, etc., creates a rigamarole of a
deployment pipeline. It often leads to increased overhead by introducing
isolation primitives (such as containers).

These and other issues may indicate systemic problems in how we design our
application for the requirements and scale we have these days.

Hypothesis
==========

Software development and bundling practices stem from the paradigm that hinges
on a single-computer view of the world. Every program has a foundation in the
form of some kind of operating system, which handles such resources as memory,
files, isolation, networking, etc. One can visualize this as a pyramid where a
program is at the top and services provided by the OS are below it. However, we
can observe that the vector of this functionality is directed towards the
management of resources of the computer that the program is running on, so it
is physically directed "inwards". This approach makes sense if your entire
system works on that computer. However, with our systems' increased complexity
and scale, this is no longer true.

What if we replaced that foundation and most of the services of such an OS were
directed outwards and provided functionality that is most useful for running
systems that span the network, can be partially rolled out, process data in
complex pipelines, etc.?

Observation: Code is smaller than data
======================================

One big issue we can observe is that the amount of data our system handle is
vastly larger than the amount of code that handles it. So why do we send the
data to the code, then? Wouldn't it make sense to send code to the data and
process it there?

A gained momentum of a virtual ISA and ABI that WebAssembly offers (while
acknowledging its flaws and immaturity) allows us to reimagine how data
processing can be done. Sandboxed code deployed across the hardware resources
that manage the data will enable us to process data in place without
serializing and transferring that data back and forth.

What if the OS, as mentioned above, provided building blocks for managing and
processing data, starting from components like B-Trees and growing into more
sophisticated data management primitives? With more sophisticated and
cluster-away scheduling, this can be a foundation for algorithms that naturally
scale across the system.

Solution: start building a micro-OS

I've given this a few tries in a few languages (like C or Zig), and this one is
done in C++. Zig is fantastic but kind of immature for where I wanted it to be
for this task.  I've done some experiments at roughly replication *some* of the
features I liked about Zig (like error management) and I am going to use these
learnings in the project.  I have slightly unorthodox (but not unheard of)
ideas about how I want to use C++ here.  I am going to ditch standard library
altogether and reimplement only those parts that I need as I see fit. I am
going to use C++20 and some gcc/clang extensions.

I've used bootboot (https://gitlab.com/bztsrc/bootboot) to avoid the burden of
booting the OS in different environments (and eventually architectures). Thank you!
I've included its mkbootimg for simplicity for th etime being

I know there are other protocols and bbootloaders (some are seemingly more
featureful, like stivale2 and limine), but I guess something simple will do for now.

The micro-OS doesn't do anything just yet, it simply boots and halts.

Going forward, what also may be interesting about this OS is that it doesn't
have to run untrusted code, as all WebAssembly modules would be either
dynamically recompiled or interpreted. Does this mean we can avoid the
separation between kernelspace and userspace? Overall, how does trusted/sandboxed
execution model change our kernel architecture.

I am using a very simple Makefile-based build system as it seems to fit best in terms
of being the easiest to bend to my needs. Dealing with CMake/Meson at this time
makes it harder than necessary, especially when I wanted to use linker without
the compiler (`-fuse-ld` issues, I am looking at you).

P.S. There are a lot of "I"s in this commit message!