Problem: no simple runtime error reporting system
Due to (known) complexities associated with C++ exceptions in freestanding
programs, ParaOS can't use them. At least, not right now.

Also, I am not very fond of return codes and the low readibility that the bring
with them.

Solution: use simple Zig-inspired errors

The idea is simple. We need a small value we can return in a global error union
in the application, so that errors can be propagated until they are handled (or
not).

This is what I've done here, in a rather simplist fashion. Obviously, no stack
traces for now (but would be great to get something going!).

To start organizing the project, I decided to use C++20 Modules. They are a bit
rough around the edges, and the implementations only starting to catch up now.
But this project is for tomorrow, so... I guess we can do this.

I've pinned C++ compiler to Clang for now (mostly because it's easier to get it
to cross-compile) as command-line options for C++20 modules haven't been
standardized yet.

Back to errors. The idea is that `Error` struct we pass around contains
a static zero-terminated string with the name of the error and to compare them (faster)
I hash them compile-time using xxHash64 (also, there's no `strcmp` yet, so that helps). So it's
128 bits per error. I am still thinking if we need to keep the hash once something
like `strcmp` is implemented). It'll be smaller if we don't keep it but error comparison
will be marginally slower. Something to think about.

Next, there's a `Result` struct that is effectively a success-tagged union of any
type and Error.

Finally, there are `tryUnwrap` macro (which uses GCC/Clang statement expression
extension) that allows to either get the value or return an error (automatic
propagation), as well as `tryCatch` macro that allows to unwrap result and if
it is an error, do something.

I was thinking how to run unit tests on `libpara.err` (oh, btw, `libpara` is
ParaOS' own stdlib, if you will. There's no standard C++ library. Go wild!),
and I didn't enjoy the idea of doing separate builds, so, instead, non-release
builds of ParaOS (that is the default) will simply run all known test cases
at a startup.

As you can tell, I've done a little bit of code organization effort in the kernel
and started isolating platform/arch-specific code to make the kernel as simple
as possible.

I've also had to add serial port functionality to display unit test results
and failures.

Things people do on long days in the maternity ward.