Problem: stack is not enough
Currently, bootboot's stack is 16K and it is not really enough
for everything ParaOS would want to achieve.

Solution: implement a rudimentary PMM
(Physical Memory Manager) as a set of simple allocators

WatermarkAllocator simply allocates chunks of contiguous memory.
ChainAllocator allows to compose allocators of the same type, and
cascade over them.

There is no freeing functionality at the moment because of its
rudimentary status.

I wanted to be able to print the amount of available memory, so I conceived
a rudimentary formatting module (libpara.formatting) and it is much nicer
to print stuff now:

```cpp
format(serial, "Available memory: ",
           this->allocator.availableMemory() / (1024 * 1024), "MB\n");
```