Problem: difficulty debugging faults
Qemu happily reboots on them, and `-d int`, while very useful, is hard to read
and not portable and won't be available on other platforms, including actual
hardware.

Solution: implement own panic handling ISR

This change required taking over the ownership of GDT and setting up the IDT,
which is a useful thing to do anyway.

I also had to implement `constexpr_loop` utility to be able to instantiate
many templates with integer parameters (interrupt handlers)