Problem: memory allocation across multiple CPUs or processes
We're not doing this but as soon as we'll initialize application
processor, we'll have to, and our current allocator (WatermarkAllocator)
won't fare well in a concurrent use scenario.

Solution: use naively implemented locking

It's not great but since we don't have a scheduler yet, there isn't
much more we need right now.

It's untested as I am not sure I can test it well before there's a scheduler,
or use of other processors actually happening. There's currently no wait to
write tests spanning multiple CPUs either.