Problem: running on SMP >2 results in stuck cores
Often 1 or 2 cores would get stuck at a random place (like a MOV instruction).

Solution: make sure we have proper stack size and usage

I noticed that it is always last 1-2 cores that would have a problem,
and sometimes one of the preceding cores would have an unexplainable
behaviour.

By analyzing and through the use of a debugger, I came to an understanding that
there's a stack overflow issue. As a knee-jerk reaction I added
a Stack Smashing Protector (ssp.cpp) but it didn't help much as it is about
stack underflow. But it's good to have it and its easy to turn it off
build-time.

Then I started reading more about how bootboot works and I realized that its
BIOS loader is actually Level 1, which means kernel core stack size is
hard-coded at 1KB and it is rather easy to blow this. BIOS loader is listed as
an obsolete loader and EFI loader should be used instead. So I extracted OVMF
images and bootboot's EFI loader and made qemu work with that instead of Legacy
BIOS. This has allowed for Level 2 and therefore 16KB stack size. Everything started
working fine.